First commit, move the plugin and the filter
